Fix : dummy  방식 수정
